+++ CHANGELOG.md
@@ -11,3 +11,4 @@ This project adheres to [Semantic Versioning](http://semver.org/).
 
 * Pass server url from localStorage to the synchronization code and connection check
 * Automatically close the panel if the add/search operations succeed
+* The error alert shows the name instead of "some error occurred" and the error message if available